feat: lass based styles for html
Introduced a concept of publishers. It's basically a theme in terms of
traditional static-site generators, but I think it might expand to more than
just a theme in future. I am not fully clear about the final requirements of
this project; but the in the light of foggy picture I have in mind, not
restricting it to established-idea of a theme feels right.