feat: footer in posts
What a mess! I need to refactor a bunch of stuff to make things consistent. I
started converting all the dom/css functions into defparameter, but then figured
changing a defun and re-evaling it with sly works, but if it is a defparameter,
I need to re-eval everything that uses it too. It is worse than macros. Or I am
too tired right now and jumped to that conclusion. Need to re-look at this with
a calmer mind.