feat: publishing tags
I decided to not have an outputs table, because for now there is a one-to-one
relationship b/w inputs and outputs. I have also removed dependencies table,
because we aren't utilizing them anywhere for now.

I have however kept a database in the mix, because the features I want to build
in future will definitely need one.

Also, rendering the tags is incomplete; styles need to be done properly. And
code need to be moved around as well.