feat: fs-provider for collecting posts from a directory
It works by itself, but it need to be changed to get it work with
org-roam-provider. We need to revisit how the rpc server is handled. We can't
have each provider close the server when it is finished. We should probably
start the rpc server before invoking any of the providers, then close it
manually when all the providers are done.