feat: category based rendering of posts
It is a bit of a hack to slap categories in fs-provider, but it works for now. I
am trying to maintain momentum and not lose it for the sake of perfection. Also,
there is a *provider-name* in elisp code, which in case of fs-provider gets
overridden on the CL side; might cause confusion in future.