feat: don't publish control-tags
I've decided to make this a feature of the clown-slick publisher. I am undecided
about the scope of clown, whether it should have a concept of control-tags or
not. The database schema for `inputs` is decided by clown, so concept of
`tags` is already there. But let's stay with the spirit of lisp, and not think
too much about future. Hack away!