refactor: major change in architecture
I think we now have enough clarity to make things a bit less hand-wavy. I have
decided to make "create a blog" limited to clown-slick, and remove all traces of
it from cl-ownpress itself.

At this point, only *org-file-provider* is working; rest of the code need to be
fixed. Although cl-ownpress itself is building, but I am certain nothing will
actually work.

I have decided to rename *fs-provider* to *org-file-provider*. Instead of trying
to make a processor which needed a lot of to-and-fro b/w rpc server. This allows
me to keep my rpc-server code untouched for now, thus reducing work.

Although docs mention existence of processors as a concept, there is no need for
them as of now.