refactor: new API for creating and using providers
Main objective of this new API is to have a per-instance name for a provider.
When the same provider instance is invoked twice, the difference b/w two
invocations (i.e content provided by first invocation but not by second) will be
deleted. This solve the problem that so far we can add and update content, but
can't delete any posts from the database which have been deleted at source.

This works for sources like notes and directory; but won't for those like
twitter. To tackle that, we'll make the purge-strategy configurable.