feat: purge-strategy for deleting provided_content
purge-strategy determines how the content is purged from the database. When nil,
provider will never purge the content. When a function, that function is called
*after* the provider has done its job. When symbol `purge-all-before-invoke`,
the provided_content added by the provider is deleted *before* it does its job.

Perhaps having a before/after hook would make more sense, but that can already
be achieved with :before/:after on invoke-provider method. I am going ahead with
this API and will let it evolve as needed.

Also,
- Set jorunal_mode to WAL for sqlite
- Enable foreign_keys pragma
- Turn all SQL tables to strict, so invalid data types will throw an error