fix: widget-children | test: widget-children
I tried a lot to figure out a way to find the children of a widget after it is
created. I want to keep the `dom-of` API; it provide a nice ergonomic way to
create and extend html of a widget. I am settling for a slightly awkward way of
figuring out children of a widget (which this commit implements). I hope I can
figure out how to make it better in future. This list of children will be used
to accumulate lass used by a widget and all of its child-widgets, without
needing the user to provide the nested widgets twice.