feat: yet another plot twist for the new widget API
Every new widget is now a class, instead of being an instance of the WIDGET
class. All this because I couldn't figure out how to walk the spinneret
WITH-HTML tree. But I think I accidentally stumbled on a nicer API than what I
had originally planned. With this API, the widget arguments are accessible to
both lass and dom (and to js in future), and even the extended dom will count
when we finally sit down to figure out what was actually rendered (e.g to
generate only required css).