refactor: move denote-provider + elisp code out to bitspook.in repo
We have a generic Emacs provider which allow running any elisp code and utilize full power of Emacs.
denote-provider relies on elisp code, and exports org files in a way which might be too specific to
my personal preferences. So I am moving it out to my website's repo. When/if it is mature enough in
future, we can move it back.