feat: introducing artifacts as an alternative to publishers
An artifact is something that will eventually be published and can depend on other artifacts.. They
provide a convenient way to represent entire "is also published" objects, which can be linked to.
This prevents the problem of "make sure X thing is published at this exact address which I am
linking to in this file"; but passing the artifact to link to and just saying "link to this
artifact, wherever it is published."

Artifacts also make the publishing API more intuitive.

Next stop: Widgets are also artifacts

Instead of *render-stack*, we can simply resolve all dependencies of a widget, if widgets are
artifacts. That will reduce the cognitive load of keeping *render-stack* healthy when converting
widget(s) to html/css. This will also makes testing much more simple, since there won't be any
filesystem side-effects to test against.