fix: replace jsonrpc with hand-rolled HTTP based RPC with clack
JSONRPC is shit. It is pretty much abandoned. Which I don't mind because CL. But out of the blue,
this weekend it has started eating up 100% CPU on my macbook (not sure if it is the case on Linux as
well). I have verified that it is not something in my code, and is a problem with JSONRPC. Although
the issue is seemingly random. I can replicate it by performing 10s of RPC calls, but not in a
controlled manner.

Since I was very disappointed with jsonrpc anyway (I stilled haven't figured out how to close the
rpcserver once started, my gh issue asking this on jsronrpc's repo is still open, unresponded); I've
decided to just move away from it. I'll make a thin facade that "there is an RPC server in
cl-ownpress" so I can switch implementations in future, and move on.